add link to view timeline after patch description
