+++ src/main/resources/messages/RadicleBundle.properties
@@ -92,5 +92,6 @@ patchProposalNoCommits=There are no commits in this patch proposal
 commits=Commits
 issueContent=This feature requires support for Issues in the rad CLI, so it will be available in a future plugin version.
 
+view.timeline=View Timeline
 patch.proposal.change.title=Change Title
 patch.proposal.submit.hint={0} to {1}
